Refactor and enhance command handling and persistence mechanisms
- Removed unused SLF4J Logger instances in QuickList and FlushAllCommand.
- Improved SubscribeCommand and UnsubscribeCommand to handle multiple subscriptions and unsubscriptions, returning responses in a more structured manner.
- Introduced MultiSuccess command result type to handle multiple responses in CommandResult.
- Updated ResponseBuilder to merge multiple ByteBuffers into a single response.
- Refactored AofRepository and RdbRepository for better error handling and code clarity, including the introduction of PersistenceException for consistent exception management.
- Enhanced ConfigurationParser to improve argument validation and error messaging.
- Simplified StreamIdComparator by converting it to an enum and cleaning up the comparison logic.
- Added BlockingException to handle blocking operation failures in a more descriptive manner.
- General code cleanup and optimization across various classes to improve readability and maintainability.

+++ src/main/java/replication/ReplicationProtocol.java
@@ -82,7 +82,7 @@ public final class ReplicationProtocol {
      * @return the size in bytes
      */
     public static long calculateCommandSize(String[] commandArgs) {
-        long totalSize = 1 + String.valueOf(commandArgs.length).length() + 2; // *<count>\r\n
+        long totalSize = 1 + (long) String.valueOf(commandArgs.length).length() + 2; // *<count>\r\n
         for (String argument : commandArgs) {
             totalSize += 1 + String.valueOf(argument.length()).length() + 2 + argument.length() + 2; // $<len>\r\n<arg>\r\n
         }
